# frontend/src/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tennis Tracker API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/players/{player_id}/season": {
      "get": {
        "summary": "Player Season Stats",
        "operationId": "player_season_stats_api_players__player_id__season_get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerSeasonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlayerProfile": {
        "properties": {
          "player_id": {
            "type": "integer",
            "title": "Player Id"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "name_acronym": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Acronym"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "country_colour": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Colour"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          }
        },
        "type": "object",
        "required": [
          "player_id"
        ],
        "title": "PlayerProfile"
      },
...
      "PlayerSeasonResponse": {
        "properties": {
          "context": {
            "$ref": "#/components/schemas/SeasonContext"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerProfile"
          },
          "summary": {
            "$ref": "#/components/schemas/PlayerSeasonSummary"
          },
          "trajectory": {
            "items": {
              "$ref": "#/components/schemas/PlayerSeasonRound"
            },
            "type": "array",
            "title": "Trajectory"
          }
        },
        "type": "object",
        "required": [
          "context",
          "player",
          "summary",
          "trajectory"
        ],
        "title": "PlayerSeasonResponse"
      },
      "PlayerSeasonRound": {
        "properties": {
          "round": {
            "type": "integer",
            "title": "Round"
          },
          "tournament_id": {
            "type": "integer",
            "title": "Tournament Id"
          },
          "tournament_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tournament Name"
          },
          "date_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Start"
          },
          "surface": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface"
          },
          "ranking_points": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ranking Points"
          },
          "ranking_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ranking Position"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "opponent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent"
          }
        },
        "type": "object",
        "required": [
          "round",
          "tournament_id"
        ],
        "title": "PlayerSeasonRound"
      },
      "PlayerSeasonSummary": {
        "properties": {
          "titles": {
            "type": "integer",
            "title": "Titles"
          },
          "grand_slams": {
            "type": "integer",
            "title": "Grand Slams"
          },
          "best_result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Result"
          },
          "best_result_tournament": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Result Tournament"
          },
          "matches_won": {
            "type": "integer",
            "title": "Matches Won"
          },
          "matches_lost": {
            "type": "integer",
            "title": "Matches Lost"
          },
          "win_pct": {
            "type": "number",
            "title": "Win Pct"
          }
        },
        "type": "object",
        "required": [
          "titles",
          "grand_slams",
          "matches_won",
          "matches_lost",
          "win_pct"
        ],
        "title": "PlayerSeasonSummary"
      },
...
      "SeasonContext": {
        "properties": {
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "tournament_id": {
            "type": "integer",
            "title": "Tournament Id"
          },
          "tournament_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tournament Name"
          },
          "tournament_official_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tournament Official Name"
          },
          "country_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Name"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "surface": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface"
          },
          "event_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Name"
          },
          "date_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Start"
          }
        },
        "type": "object",
        "required": [
          "year",
          "tournament_id"
        ],
        "title": "SeasonContext"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}